+++ lib/src/tokens/token_drawer_theme.dart
@@ -1,7 +1,7 @@
-import 'package:flutter/material.dart';
+import 'package:design_sos/design_sos.dart';
 import 'package:design_sos/src/utils/border_radius_cast.dart';
 import 'package:design_sos/src/utils/color_conversion.dart';
-import 'package:design_sos/design_sos.dart';
+import 'package:flutter/material.dart';
 
 class TokenDrawerTheme extends DrawerThemeData with TokenAttributes {
   TokenDrawerTheme({required Color super.backgroundColor, super.scrimColor, super.elevation, super.width, BorderRadius? borderRadius})
@@ -10,7 +10,7 @@ class TokenDrawerTheme extends DrawerThemeData with TokenAttributes {
   factory TokenDrawerTheme.fromJson(Map<String, dynamic> json) => TokenDrawerTheme(
       backgroundColor: convertColor(json['backgroundColor'] ?? '') ?? Colors.white,
       scrimColor: convertColor(json['scrimColor'] ?? ''),
-      elevation: json['elevation'] ?? 16,
+      elevation: json['elevation'].toDouble() ?? (16).toDouble(),
       width: json['width'] != 0 ? json['width'] : null,
       borderRadius: borderRadiusFromJson(json['borderRadius'] ?? {}));
